+++ packages/frontend/src/pages/users/index.vue
@@ -20,6 +20,7 @@
         <v-row>
           <v-col md="3">
             <v-list-item-title :data-testid="'test-'+user.username">
+              <user-avatar :user="user" class="mr-2" />
               {{ user.username }}
               <template v-if="user.name">
                 ({{ user.name }})